requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-phonenumber-field==8.4.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.6
markupsafe==3.0.3
openapi-codec==1.3.2
packaging==25.0
phonenumbers==9.0.19
psycopg2-binary==2.9.11
pyjwt==2.10.1
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
simplejson==3.20.2
sqlparse==0.5.4
uritemplate==4.2.0
urllib3==2.5.0
